[tests] Fix compilation issue if parent project use C++ 17
The library require at least a C++ 17 compiler. Sadly, designated initializers in C++ were introduced only in C++20 (in C, that was in C99 !).

Designated initializers are useful in unit-tests and requiring the library to be build by a C++ 20 compiler at least, would be too much.

So, keep the minimal standard to C++ 17 but compile unit-tests with C++ 20 minimum.